Squashed commit of the following
commit 51f43023 
Author: patricia <patriciapavia236@gmail.com> 
Date: Thu Aug 27 2020 10:21:03 GMT+1000 (Australian Eastern Standard Time) 

    LWA-670 fixed prior error


commit a6e60ea4 
Author: patricia <patriciapavia236@gmail.com> 
Date: Thu Aug 27 2020 10:07:23 GMT+1000 (Australian Eastern Standard Time) 

    LWA-670 fixed prior error


commit d6f0db27 
Author: patricia <patriciapavia236@gmail.com> 
Date: Wed Aug 26 2020 16:54:02 GMT+1000 (Australian Eastern Standard Time) 

    LWA-670


commit 9e0e888c 
Author: patricia <patriciapavia236@gmail.com> 
Date: Wed Aug 26 2020 16:48:03 GMT+1000 (Australian Eastern Standard Time) 

    LWA-670

+++ src/controller/BannerController.ts
@@ -1,4 +1,4 @@
-import {Response} from "express";
+import { Response, response } from "express";
 import {
     Get,
     JsonController,
@@ -12,9 +12,9 @@ import {
     Res
 } from "routing-controllers";
 
-import {BaseController} from "./BaseController";
-import {Banner} from "../models/Banner";
-import {isPhoto, fileExt, timestamp} from "../utils/Utils";
+import { BaseController } from "./BaseController";
+import { Banner } from "../models/Banner";
+import { isPhoto, fileExt, timestamp, isNotNullAndUndefined } from "../utils/Utils";
 
 @JsonController("/banners")
 export class BannerController extends BaseController {
@@ -85,8 +85,23 @@ export class BannerController extends BaseController {
 
     @Authorized()
     @Delete('/id/:id')
-    async deleteById(@Param("id") id: number) {
-        return this.bannerService.deleteById(id);
-    }
+    async deleteById(
+        @Param("id") id: number,
+        @Res() response: Response) {
+        const banner = await this.bannerService.findById(id);
 
+        if (isNotNullAndUndefined(banner)) {
+
+            const fileName = await this.firebaseService.getFileNameFromUrl(JSON.stringify(banner.bannerUrl), 'banner%2F');
+            await this.firebaseService.removeMedia(fileName)
+            return this.bannerService.deleteById(id);
+
+        } else {
+
+            return response.status(400).send({
+                name: 'delete_error',
+                message: `Banner was already deleted`
+            });
+        }
+    }
 }
